Moved all initialization into host_setup.cpp
